refactored message handler into a seperate module
